Use common error handling for all cli tools
* DeviceExceptions are catched in the base group (ExceptionHandlerGroup)
* Token & ip validation is moved to click_common module
