+++ miio/click_common.py
@@ -1,0 +1,35 @@
+"""Click commons.
+
+This file contains common functions for cli tools.
+"""
+import click
+import ipaddress
+import miio
+import logging
+
+
+_LOGGER = logging.getLogger(__name__)
+
+
+def validate_ip(ctx, param, value):
+    try:
+        ipaddress.ip_address(value)
+        return value
+    except ValueError as ex:
+        raise click.BadParameter("Invalid IP: %s" % ex)
+
+
+def validate_token(ctx, param, value):
+    token_len = len(value)
+    if token_len != 32:
+        raise click.BadParameter("Token length != 32 chars: %s" % token_len)
+    return value
+
+
+class ExceptionHandlerGroup(click.Group):
+    def __call__(self, *args, **kwargs):
+        try:
+            return self.main(*args, **kwargs)
+        except miio.DeviceException as ex:
+            _LOGGER.debug("Exception: %s", ex, exc_info=True)
+            click.echo(click.style("Error: %s" % ex, fg='red', bold=True))
