+++ miio/vacuum_cli.py
@@ -5,7 +5,6 @@ import pretty_cron
 import ast
 import sys
 import json
-import ipaddress
 import time
 import pathlib
 from appdirs import user_cache_dir
@@ -19,31 +18,14 @@ if sys.version_info < (3, 4):
     sys.exit(1)
 
 import miio  # noqa: E402
+from miio.click_common import (ExceptionHandlerGroup, validate_ip,
+                               validate_token)
 
 _LOGGER = logging.getLogger(__name__)
 pass_dev = click.make_pass_decorator(miio.Device, ensure=True)
 
 
-def validate_ip(ctx, param, value):
-    if value is None:
-        return value
-    try:
-        ipaddress.ip_address(value)
-        return value
-    except ValueError as ex:
-        raise click.BadParameter("Invalid IP: %s" % ex)
-
-
-def validate_token(ctx, param, value):
-    if value is None:
-        return value
-    token_len = len(value)
-    if token_len != 32:
-        raise click.BadParameter("Token length != 32 chars: %s" % token_len)
-    return value
-
-
-@click.group(invoke_without_command=True)
+@click.group(invoke_without_command=True, cls=ExceptionHandlerGroup)
 @click.option('--ip', envvar="MIROBO_IP", callback=validate_ip)
 @click.option('--token', envvar="MIROBO_TOKEN", callback=validate_token)
 @click.option('-d', '--debug', default=False, count=True)
